Added interfaces along with implementation for some services

+++ cryptoCurrency/src/main/java/org/example/cryptocurrency/repository/CryptoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CryptoRepository extends JpaRepository<Crypto, Long> {
+public interface CryptoRepository extends JpaRepository<Crypto, Integer> {
     Optional<Crypto> findBySymbol(String symbol);
 }
