+++ cryptoCurrency/src/main/java/org/example/cryptocurrency/repository/UserAccountRepository.java
@@ -3,6 +3,6 @@ package org.example.cryptocurrency.repository;
 import org.example.cryptocurrency.models.UserAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
+public interface UserAccountRepository extends JpaRepository<UserAccount, Integer> {
 
 }
